#1325 protocol.rs split into protocol/ module
Split userspace-dp/src/protocol.rs (3042 LOC) into a protocol/
directory module with 7 domain files plus tests + slim mod.rs,
following the Wave-1 module/foo.rs layout convention. Pure code
motion. Wire format byte-identical. Public API unchanged.

Layout:
- protocol/mod.rs       — module decls + pub(crate) use glob
                          re-exports for back-compat
- protocol/snapshot.rs  — ConfigSnapshot + interface/route/zone/
                          fabric/tunnel/neighbor/mirror snapshots
                          + custom Debug for TunnelEndpointSnapshot
                          + MapPins, UserspaceCapabilities,
                          FlowSnapshot (13 named items)
- protocol/cos.rs       — ClassOfServiceSnapshot tree (9 config
                          DTOs) + CoSInterfaceStatus, CoSQueueStatus,
                          CoSActiveFlowCountStatus (14 named items)
- protocol/nat.rs       — Source/Destination/Static NAT, NAT64,
                          Nptv6 + SourceNatPoolStatus (6 items)
- protocol/security.rs  — Screen/firewall-filter/policer/three-color
                          policer/flow-export/policy-rule snapshots
                          + their counter status twins (11 items)
- protocol/control.rs   — protocol-version consts + ControlRequest/
                          Response, ProcessStatus, session-sync
                          wire shapes (14 items, deepest module —
                          depends on every leaf + snapshot)
- protocol/binding.rs   — BindingStatus (~420 LOC), BindingCounters-
                          Snapshot + From impl + _ASSERT compile-
                          time invariant, HAGroupStatus, QueueStatus,
                          WorkerRuntimeStatus, ExceptionStatus,
                          SessionDeltaInfo, u64_is_zero (8 items)
- protocol/resolution.rs — PacketResolution + FlowTupleStatus
                          (+ from_session_key impl) + FlowWorkerStatus
                          (3 items)
- protocol/tests.rs     — 631 LOC of inline serde round-trip /
                          backward-compat tests, relocated as a
                          single cohesive submodule since every
                          test exercises ProcessStatus that
                          aggregates leaves from every domain.

Total: 66 structs + 2 consts + 1 fn = 69 named items (matches
pre-split count).

u64_is_zero references switched to the absolute path
`crate::protocol::u64_is_zero` in the
`#[serde(skip_serializing_if = ...)]` attribute on
HAGroupStatus.lease_until, per Codex r1/r2's verified
serde_derive semantics (the string is parsed as syn::ExprPath
and emitted directly into the generated code, so absolute paths
resolve regardless of submodule placement).

Zero call-site edits: the 5 `use crate::protocol::X` sites
and 12 `crate::protocol::T` path callers across the crate all
resolve through mod.rs's glob re-exports unchanged. main.rs's
`use protocol::*;` still surfaces every type and the two
constants at crate root for downstream consumers like
afxdp/coordinator/tests.rs.

Plan + adversarial review:
- docs/pr/1325-protocol-split/plan.md (v3 PLAN-READY-pending)
- Codex r1 PLAN-KILL → r2 PLAN-NEEDS-MAJOR → v3 folds
- AGY r1 PLAN-NEEDS-MINOR → r2 PLAN-NEEDS-MINOR → v3 folds

Tests:
- cargo build --release --tests clean
- cargo test --release --bin xpf-userspace-dp:
  1417 passed / 0 failed / 2 ignored (was 1395 pre-split because
  the relocated #[cfg(test)] mod tests now lives in the same
  test binary instead of an external integration test)
- 5/5 flake check on the 22 protocol tests
- Go suite clean
- Pre-existing snat_contract_doc_guard failure on both master
  and this branch (unrelated to refactor)

Closes #1325
