+++ docs/pr/1325-protocol-split/reviewer-ids.md
@@ -6,11 +6,13 @@ For continuation across companion-session loss. Append as IDs are dispatched.
 
 ### Codex
 
-- *(pending dispatch)*
+- r1: task-mpmuqtzp-jmkjd9 (plan @ 39df2e52) — PLAN-KILL, 5 fixable
+- r2: task-mpmv2opw-ml9970 (plan @ 592ce725) — pending
 
 ### AGY
 
-- *(pending dispatch)*
+- r1: adversarial-review-mpmuvd0z-7um75c (plan @ 39df2e52) — PLAN-NEEDS-MINOR, 4 items
+- r2: adversarial-review-mpmv2wym-bzqr57 (plan @ 592ce725) — pending
 
 ## Code review
